Update to login failed error messages
Correct me if I am wrong here.

But in _authenticateUserFromRequest() function from line 38
Isn't email always undefined as we don't have a function login function in any SDK which uses email for login.

I have updated the messages assuming the above state is incorrect.

I believe that the handling of error messages and error codes needs to be updated. Right?